Aplicando conceitos da Mentoria #6 junto ao REST da Mentoria #5

+++ src/main/java/com/avanade/dio/api/controllers/LojaController.java
@@ -44,7 +44,7 @@ public class LojaController {
     @DeleteMapping
     @RequestMapping("{id}")
     @ResponseStatus(HttpStatus.ACCEPTED)
-    public void excluir(@PathVariable Long id){
+    public void excluir(@PathVariable Integer id){
         lojaService.excluir(id);
     }    
 }
